Replace wxPython data viewer with the PyQt version
